extracting voteRow to a component
