+++ apps/davi/src/components/ProposalVotesCard/components/index.ts
@@ -1,0 +1,1 @@
+export { default as VoteRow } from './VoteRow';
